pyproject.toml
[project]
name = "agile-display"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.2.0",
    "requests>=2.33.1",
]
